fix: validate market interval and support large order book levels

+++ exchange-simulator/Models/TradingCore/OrderBookLevel.cs
@@ -1,7 +1,9 @@
+using System.Numerics;
+
 namespace exchange_simulator.Models.TradingCore;
 
 public sealed record OrderBookLevel(
     decimal Price,
-    long Size
+    BigInteger Size
     
     );
